+++ fotoobo/cli/convert.py
@@ -4,13 +4,14 @@ The fotoobo convert commands
 
 import logging
 from pathlib import Path
+from typing import Annotated
 
 import typer
-from typing_extensions import Annotated
 
 from fotoobo.helpers import cli_path
 from fotoobo.helpers.files import load_json_file, save_json_file
 from fotoobo.tools import convert
+
 
 app = typer.Typer(no_args_is_help=True, rich_markup_mode="rich")
 log = logging.getLogger("fotoobo")