+++ fotoobo/cli/fgt/get/cmdb/firewall/firewall.py
@@ -5,7 +5,7 @@ The FortiGate commands
 import logging
 
 import typer
-from typing_extensions import Annotated
+from typing import Annotated
 
 from fotoobo.exceptions import GeneralError
 from fotoobo.helpers import cli_path