+++ fotoobo/cli/get.py
@@ -7,7 +7,7 @@ import logging
 import typer
 from rich import print as rich_print
 from rich.panel import Panel
-from typing_extensions import Annotated
+from typing import Annotated
 
 from fotoobo.helpers import cli_path
 from fotoobo.tools import get