+++ WHATSNEW.md
@@ -1,4 +1,5 @@
 ### Changed
 
-- Update pyproject.toml to new format an remove deprecated settings
+- Update pyproject.toml to new format and remove deprecated settings
 - Update GitHub action for new pyproject.toml format (use --extras for optional dependencies)
+- Change the way how to import `Annotated` (from `typing`)